+++ CHANGES.rst
@@ -5,7 +5,7 @@
 1.13 (unreleased)
 =================
 
-- Nothing changed yet.
+- Update test infrastructure to `icemac.addressbook >= 3.0`.
 
 
 1.12 (2017-01-21)